pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tdf-data-bridge"
version = "1.0.0"
description = "ProForm TDF Bike Control via ANT+ and BLE FTMS"
readme = "README.md"
license = "MIT"
authors = [
  { name = "TheItalianDataGuy" }
]
dependencies = [
  "bleak==0.22.3",
  "openant==1.3.3",
  "pyserial==3.5"
]

[project.scripts]
tdf-bridge = "tdf_data_bridge:main"
